+++ index.js
@@ -1,11 +1,12 @@
 module.exports = {
-  Switch: require("./Switch"),
-  ColorPicker: require("./ColorPicker"),
-  YoutubeVideo: require("./YoutubeVideo"),
-  VimeoVideo: require("./VimeoVideo"),
-  LoadingSpinner: require("./LoadingSpinner"),
   Button: require("./Button"),
+  ColorPicker: require("./ColorPicker"),
+  LoadingSpinner: require("./LoadingSpinner"),
   Pagination: require("./Pagination"),
   Panel: require("./Panel"),
   ProgressBar: require("./ProgressBar"),
+  Switch: require("./Switch"),
+  Tabs: require("./Tabs"),
+  YoutubeVideo: require("./YoutubeVideo"),
+  VimeoVideo: require("./VimeoVideo"),
 };