+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Removed MediaObject and FloatingLabelInput components.
 * Switched CSS pre-processor from Less to Sass.
 * Switched from grunt-react to grunt-babel.
-* Updated React dependency to v15.0.1.
+* Updated React to v15.0.1.
 * Switch is now a controlled component. Previously it was uncontrolled, maintaining it's own state.
 * Added LoadingSpinner component
 * Removed TagsInput. [React Select](http://jedwatson.github.io/react-select/) is a better alternative.